requirements.txt
joblib==1.2.0
joblib==1.3.2
matplotlib==3.6.2
matplotlib==3.8.2
numpy==1.23.4
pandas==1.5.1
Pillow==9.3.0
Pillow==10.4.0
scikit_learn==1.4.0
seaborn==0.13.2
streamlit==1.30.0
